instructions for use
Step 1: Put the "Whitelist Tablet" prefab in your scene

Step 2: unpack the prefab completely and remove the white box.

step 3: move the white box to wherever you want to be whitelist access only and resize. whatever that box covers will only be able to be accessed by whitelisted users.

step 4: make sure the whitelist is connected to the collider script (on the white box) and the tablets' whitelist manager. 

step 5: place the tablet wherever you want and test it out.

note: the whitelist in the tablet throws an error saying it needs an upgrade. Just ignore that and delete the component that is saying that.